.claude/worktrees/goofy-torvalds-d78597/requirements.txt
qdrant-client>=1.12.0
sentence-transformers>=2.2.0
python-dotenv>=1.0.0
tqdm>=4.65.0
anthropic>=0.40.0
cohere>=5.0.0
ragie>=1.0.0
openai>=1.50.0
google-genai>=1.73.1
pyyaml>=6.0
matplotlib>=3.7.0
